+++ packages/core-js-compat/helpers.js
@@ -13,7 +13,9 @@ class SemVer {
     this.minor = $minor ? +$minor : 0;
     this.patch = $patch ? +$patch : 0;
   }
-  toString() {
+  toString(slice) {
+    if (slice === 'major') return `${ this.major }`;
+    if (slice === 'minor') return `${ this.major }.${ this.minor }`;
     return `${ this.major }.${ this.minor }.${ this.patch }`;
   }
 }